backend/requirements.txt
# Web框架
Flask==2.3.3
flask-socketio==5.3.6
Werkzeug==2.3.7

# 数据处理
pandas==2.1.1
openpyxl==3.0.9
xlrd==2.0.1

# HTTP请求
requests==2.26.0

# WebSocket支持
python-socketio==5.9.0
python-engineio==4.8.0
simple-websocket==0.10.1

# 环境变量管理
python-dotenv==0.19.0